Draft `Panes` widget and `panes` example
