+++ Cargo.toml
@@ -44,6 +44,7 @@ members = [
     "examples/events",
     "examples/geometry",
     "examples/integration",
+    "examples/panes",
     "examples/pokedex",
     "examples/progress_bar",
     "examples/solar_system",
