+++ examples/clock/Cargo.toml
@@ -5,9 +5,6 @@ authors = ["Héctor Ramón Jiménez <hector0193@gmail.com>"]
 edition = "2018"
 publish = false
 
-[features]
-canvas = []
-
 [dependencies]
 iced = { path = "../..", features = ["canvas", "async-std", "debug"] }
 iced_native = { path = "../../native" }
