+++ examples/clock/README.md
@@ -2,7 +2,7 @@
 
 An application that uses the `Canvas` widget to draw a clock and its hands to display the current time.
 
-The __[`main`]__ file contains all the code of the example.
+The __[`lib`]__ file contains the relevant code of the example.
 
 <div align="center">
   <img src="https://user-images.githubusercontent.com/518289/74716344-a3e6b300-522e-11ea-8aea-3cc0a5100a2e.gif">
@@ -13,4 +13,4 @@ You can run it with `cargo run`:
 cargo run --package clock
 ```
 
-[`main`]: src/main.rs
+[`lib`]: src/lib.rs
