+++ examples/panes/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "panes"
+version = "0.1.0"
+authors = ["Héctor Ramón Jiménez <hector0193@gmail.com>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+iced = { path = "../..", features = ["async-std"] }
+clock = { path = "../clock" }
+stopwatch = { path = "../stopwatch" }
