+++ native/src/element.rs
@@ -243,7 +243,7 @@ where
     }
 
     /// Computes the _layout_ hash of the [`Element`].
-    /// 
+    ///
     /// [`Element`]: struct.Element.html
     pub fn hash_layout(&self, state: &mut Hasher) {
         self.widget.hash_layout(state);
