+++ wgpu/src/renderer/widget.rs
@@ -2,6 +2,7 @@ mod button;
 mod checkbox;
 mod column;
 mod container;
+mod panes;
 mod progress_bar;
 mod radio;
 mod row;
